+++ lib/macro/macro-tab.cpp
@@ -5,7 +5,9 @@
 #include "macro-properties.hpp"
 #include "macro-tree.hpp"
 #include "macro.hpp"
+#include "math-helpers.hpp"
 #include "name-dialog.hpp"
+#include "path-helpers.hpp"
 #include "switcher-data.hpp"
 #include "ui-helpers.hpp"
 #include "utility.hpp"
