+++ lib/utils/export-symbol-helper.hpp
@@ -1,5 +1,12 @@
 #pragma once
 
+#ifdef UNIT_TEST
+
+#define EXPORT
+#define ADVSS_EXPORT
+
+#else
+
 #ifdef _MSC_VER
 #define EXPORT __declspec(dllexport)
 #else
@@ -12,3 +19,5 @@
 #else
 #define ADVSS_EXPORT Q_DECL_IMPORT
 #endif
+
+#endif // UNIT_TEST
