+++ lib/utils/math-helpers.hpp
@@ -9,8 +9,9 @@ namespace advss {
 
 std::variant<double, std::string>
 EvalMathExpression(const std::string &expression);
-bool IsValidNumber(const std::string &);
-EXPORT std::optional<double> GetDouble(const std::string &);
-EXPORT std::optional<int> GetInt(const std::string &);
+bool IsValidNumber(const std::string &str);
+EXPORT std::optional<double> GetDouble(const std::string &str);
+EXPORT std::optional<int> GetInt(const std::string &str);
+EXPORT bool DoubleEquals(double left, double right, double epsilon);
 
 } // namespace advss
