+++ plugins/base/macro-condition-obs-stats.cpp
@@ -1,8 +1,9 @@
 #include "macro-condition-obs-stats.hpp"
 #include "layout-helpers.hpp"
-#include "utility.hpp"
+#include "math-helpers.hpp"
 
 #include <obs-frontend-api.h>
+#include <QListView>
 
 namespace advss {
